Moved soundOverrides from preview examples to preview as the UI Input module prefab is dependent on it
